refactor: split ArcCanvas.jsx into per-component files (#2835)
