+++ client/src/components/pipeline/arcCanvas/shared.js
@@ -1,0 +1,8 @@
+// Shared design tokens for the Arc Canvas subtree. Severity → Tailwind class
+// map used by both verification panels (VerifyResults) and the manuscript-
+// completeness panel (CompletenessResults).
+export const SEVERITY_COLORS = {
+  high: 'text-port-error border-port-error/40 bg-port-error/10',
+  medium: 'text-port-warning border-port-warning/40 bg-port-warning/10',
+  low: 'text-gray-400 border-gray-500/30 bg-gray-700/20',
+};
